i16 Evaluation and Lockless Transposition Table

+++ src/bm/bm_runner/ab_consts.rs
@@ -4,8 +4,8 @@ pub const THREAT_MOVE_CNT: usize = 1;
 pub const IID_DEPTH: u32 = 5;
 
 pub const F_PRUNE_DEPTH: u32 = 6;
-pub const F_PRUNE_THRESHOLD_BASE: i32 = 200;
-pub const F_PRUNE_THRESHOLD_FACTOR: i32 = 200;
+pub const F_PRUNE_THRESHOLD_BASE: i16 = 200;
+pub const F_PRUNE_THRESHOLD_FACTOR: i16 = 200;
 pub const DO_F_PRUNE: bool = true;
 
 pub const NULL_MOVE_REDUCTION_BASE: u32 = 3;
@@ -30,12 +30,12 @@ pub const LMP_FACTOR: f32 = 0.88;
 pub const DO_LMP: bool = false;
 
 pub const QUIESCENCE_SEARCH_DEPTH: u32 = 30;
-pub const DELTA_MARGIN: i32 = 1000;
+pub const DELTA_MARGIN: i16 = 1000;
 pub const DO_DELTA_PRUNE: bool = true;
 pub const DO_SEE_PRUNE: bool = true;
 
 pub const FAIL_CNT: u8 = 10;
-pub const WINDOW_START: i32 = 25;
-pub const WINDOW_FACTOR: i32 = 1;
-pub const WINDOW_DIVISOR: i32 = 4;
-pub const WINDOW_ADD: i32 = 5;
+pub const WINDOW_START: i16 = 25;
+pub const WINDOW_FACTOR: i16 = 1;
+pub const WINDOW_DIVISOR: i16 = 4;
+pub const WINDOW_ADD: i16 = 5;
